+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Pass execute arguments to operation `Closure`
+
 ## [0.4.0] - 2020-10-26
 
 - Added Exception catching to insistent operations